refactor(eventstream): rename transformStreamController to streamController
