+++ packages/eventstream/src/streamController.ts
